refactor(Application): change types import dir, and clean comments
