--- src/components/defintions.components.ts
@@ -1,2 +1,0 @@
-export type TSizes = 'xs' | 'sm' | 'md' | 'lg' | 'xl';
-export type TVariants = 'primary' | 'info' | 'warning' | 'success' | 'danger';
