+++ src/features/tickets/hooks/use-get-tickets.ts
@@ -1,5 +1,5 @@
 import useGetAllData from '../../../hooks/use-get-all-data'
-import { ITicket } from '../types'
+import { ITicket } from '../../../types/types'
 
 export default function useGetTickets() {
   return useGetAllData<ITicket[]>('/tickets')
