+++ src/features/users/hooks/use-get-single-user.ts
@@ -1,5 +1,5 @@
 import useGetSingleApiData from '../../../hooks/use-get-single-api-data'
-import { IUser } from '../types'
+import { IUser } from '../../../types/types'
 
 export default function useGetSingleUser(userId: string = "") {
   return useGetSingleApiData<IUser>({ endpoint: '/users', id: userId })
