+++ src/features/users/hooks/use-get-users.ts
@@ -1,5 +1,5 @@
 import useGetAllData from '../../../hooks/use-get-all-data'
-import { IUser } from '../types'
+import { IUser } from '../../../types/types'
 
 export default function useGetUsers() {
   return useGetAllData<IUser[]>('/users')
