+++ src/features/users/schemas/user.schema.ts
@@ -1,6 +1,5 @@
 import { z } from "zod";
 import { USER_ROLES_VALUES } from "../../../constraints/constraints";
-// import { USER_ROLES_VALUES } from "../constraints/constraints";
 
 const userSchema = {
   name: z.string().min(3, 'name should be at least 3 characters').max(20, 'name should be maximum 20 characters '),
