--- src/features/users/types.ts
@@ -1,10 +1,0 @@
-export type Role = 'admin' | 'manager' | 'csr' | 'view_only'
-
-export interface IUser {
-  _id: string,
-  name: string,
-  email: string,
-  role: Role,
-  createdAt?: string,
-  updatedAt?: string,
-}
